a serious new function _cylinder()
it slices, it dices. ok, all it is is a cylinder, but it does everything
you want it to. it defaults to using polyholes, because other
calculations are pointless. it also uses length to create a slot. and it
can take any combination of r,r1,d2 etc and act exactly like a normal
cylinder. and you can center it any way you want.
there is a helper function _cylinder_fn which will return the number of
sides that _cylinder will use.
it creates holes that are symmetrical in all ways. it's awesome, and
should be used for everything.